+++ chapter-2/chapter-2.rb
@@ -1,0 +1,33 @@
+puts 
+puts "# = = = = = = = = = = "
+puts "# Chapter 2"
+puts "# File: chapter-2.rb"
+puts "# = = = = = = = = = = "
+puts
+
+puts 'Letters:'
+puts 'Hello. Ruby rocks.'
+puts '5 is my favorite number... what is yours?'
+puts 'Snoopy says #%^?&*@! when he stubs his toe.'
+puts ''
+
+puts 'String Arithmetic:'
+puts 'I like ' + 'apple pie.'
+puts 'blink ' * 4
+
+puts 'Numbers vs Strings'
+puts  12  +  12
+puts '12' + '12'
+puts '12  +  12'
+puts  2  *  5
+puts '2' *  5
+puts '2  *  5'
+puts ''
+
+puts 'Problems:'
+puts 'You\'re swell!'
+puts 'You\'re swell!'
+puts 'backslash at the end of a string:  \\'
+puts 'up\\down'
+puts 'up\down'
+puts ''